+++ src/Umbraco.Infrastructure/Persistence/Repositories/Implement/LanguageRepository.cs
@@ -391,7 +391,7 @@ internal sealed class LanguageRepository : AsyncEntityRepositoryBase<int, ILangu
             await db.Language
                 .Where(x => x.FallbackLanguageId == entity.Id)
                 .ExecuteUpdateAsync(setter => setter
-                    .SetProperty(x => x.FallbackLanguageId, (int?)null));
+                    .SetProperty(x => x.FallbackLanguageId, (short?)null));
 
             // delete
             await base.PersistDeletedItemAsync(entity);
@@ -484,13 +484,13 @@ internal sealed class LanguageRepository : AsyncEntityRepositoryBase<int, ILangu
         }
     }
 
-    private int? GetFallbackLanguageId(ILanguage entity)
+    private short? GetFallbackLanguageId(ILanguage entity)
     {
-        int? fallbackLanguageId = null;
+        short? fallbackLanguageId = null;
         if (entity.FallbackIsoCode.IsNullOrWhiteSpace() == false &&
             _codeIdMap.TryGetValue(entity.FallbackIsoCode, out var languageId))
         {
-            fallbackLanguageId = languageId;
+            fallbackLanguageId = (short)languageId;
         }
 
         return fallbackLanguageId;
